Delete IdempotentRequest and refactor MPRequest
