+++ src/main/java/com/mercadopago/client/customer/CustomerCardClient.java
@@ -2,12 +2,12 @@ package com.mercadopago.client.customer;
 
 import com.google.gson.JsonObject;
 import com.mercadopago.MercadoPagoConfig;
-import com.mercadopago.client.IdempotentRequest;
 import com.mercadopago.client.MercadoPagoClient;
 import com.mercadopago.core.MPRequestOptions;
 import com.mercadopago.exceptions.MPException;
 import com.mercadopago.net.HttpMethod;
 import com.mercadopago.net.MPHttpClient;
+import com.mercadopago.net.MPRequest;
 import com.mercadopago.net.MPResourceList;
 import com.mercadopago.net.MPResponse;
 import com.mercadopago.resources.customer.CustomerCard;
@@ -92,14 +92,14 @@ public class CustomerCardClient extends MercadoPagoClient {
       String customerId, CustomerCardCreateRequest request, MPRequestOptions requestOptions)
       throws MPException {
     JsonObject payload = Serializer.serializeToJson(request);
-    IdempotentRequest idempotentRequest =
-        IdempotentRequest.buildRequest(
+    MPRequest mpRequest =
+        MPRequest.buildRequest(
             String.format("/v1/customers/%s/cards", customerId),
             HttpMethod.POST,
             payload,
             null,
             requestOptions);
-    MPResponse response = send(idempotentRequest);
+    MPResponse response = send(mpRequest);
 
     CustomerCard card = Serializer.deserializeFromJson(CustomerCard.class, response.getContent());
     card.setResponse(response);
