+++ src/main/java/com/mercadopago/net/MPRequest.java
@@ -1,15 +1,16 @@
 package com.mercadopago.net;
 
 import com.google.gson.JsonObject;
+import com.mercadopago.core.MPRequestOptions;
 import java.util.Map;
-import lombok.AllArgsConstructor;
+import java.util.Objects;
+import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
 
 /** MPRequest class. */
 @Getter
 @Builder
-@AllArgsConstructor
 public class MPRequest {
   private final String uri;
 
@@ -28,4 +29,52 @@ public class MPRequest {
   private final int connectionRequestTimeout;
 
   private final int socketTimeout;
+
+  /**
+   * Method responsible for build MP request.
+   *
+   * @param path path
+   * @param method method
+   * @param payload payload
+   * @param queryParams queryParams
+   * @param requestOptions requestOptions
+   * @return MPRequest
+   */
+  public static MPRequest buildRequest(
+      String path,
+      HttpMethod method,
+      JsonObject payload,
+      Map<String, Object> queryParams,
+      MPRequestOptions requestOptions) {
+    MPRequest mpRequest;
+    String uri = UrlFormatter.format(path);
+
+    if (Objects.nonNull(requestOptions)) {
+      mpRequest =
+          MPRequest.builder()
+              .uri(uri)
+              .method(method)
+              .headers(requestOptions.getCustomHeaders())
+              .payload(payload)
+              .queryParams(queryParams)
+              .accessToken(requestOptions.getAccessToken())
+              .connectionRequestTimeout(requestOptions.getConnectionRequestTimeout())
+              .connectionTimeout(requestOptions.getConnectionTimeout())
+              .socketTimeout(requestOptions.getSocketTimeout())
+              .build();
+    } else {
+      mpRequest = MPRequest.builder().uri(uri).method(method).payload(payload).build();
+    }
+
+    return mpRequest;
+  }
+
+  /**
+   * Method responsible for create a new Idempotency key.
+   *
+   * @return Idempotency key
+   */
+  public String createIdempotencyKey() {
+    return UUID.randomUUID().toString();
+  }
 }
