+++ src/test/java/mercadopago/client/MercadoPagoClientTest.java
@@ -10,7 +10,6 @@ import static org.mockito.Mockito.verify;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.mercadopago.client.IdempotentRequest;
 import com.mercadopago.client.MercadoPagoClient;
 import com.mercadopago.core.MPRequestOptions;
 import com.mercadopago.exceptions.MPApiException;
@@ -62,26 +61,6 @@ public class MercadoPagoClientTest extends BaseClientTest {
         .execute(any(HttpRequestBase.class), any(HttpContext.class));
     MPRequest mpRequest =
         MPRequest.builder()
-            .uri("https://test.com")
-            .method(HttpMethod.POST)
-            .payload(requestObject)
-            .build();
-    MPResponse mpResponse = testClient.sendRequest(mpRequest);
-
-    assertNotNull(mpResponse);
-    assertEquals(200, (int) mpResponse.getStatusCode());
-  }
-
-  @Test
-  public void sendIdempotentRequestWithBodySuccess() throws IOException, MPException {
-    String request = MockHelper.readRequestFile(requestFile);
-    JsonObject requestObject = JsonParser.parseString(request).getAsJsonObject();
-    HttpResponse httpResponse = MockHelper.generateHttpResponseFromFile(responseFile, 200);
-    doReturn(httpResponse)
-        .when(httpClientMock)
-        .execute(any(HttpRequestBase.class), any(HttpContext.class));
-    MPRequest mpRequest =
-        IdempotentRequest.idempotentBuilder()
             .uri("https://test.com")
             .method(HttpMethod.POST)
             .payload(requestObject)
